+++ IntroDiagrammes/src/version2/Account.java
@@ -4,18 +4,18 @@ package version2;
 class Account {
 
     private String accountNumber;
-    private float balance;
+    private double balance;
 
-    Account(String accountNumber, float balance) {
+    Account(String accountNumber, double balance) {
         this.accountNumber = accountNumber;
         this.balance = balance;
     }
 
-    void credit(float amount) {
+    void credit(double amount) {
         balance = balance + amount;
     }
 
-    void debit(float amount) {
+    void debit(double amount) {
         balance = balance - amount;
     }
